requirements.txt
# Core dependencies
altair==5.5.0
anyascii==0.3.2
anyio==4.2.0
argon2-cffi==21.3.0
arrow==1.3.0
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
bottleneck==1.3.7
brotli==1.1.0
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
emoji==2.14.1
en-core-web-sm==3.8.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
fasttext==0.9.3
filelock==3.16.1
fonttools==4.55.6
fqdn==1.5.1
fsspec==2024.12.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==0.18.1
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
ipykernel==6.29.3
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
json5==0.9.14
jsonpointer==3.0.0
jsonschema==4.21.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyterlab==4.1.5
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
llvmlite==0.44.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.1.2
numba==0.61.0
numexpr==2.10.0
numpy==1.26.4
openai==1.60.1
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
Pillow==11.1.0
platformdirs==4.2.1
ply==3.11
preshed==3.0.9
prometheus-client==0.20.0
protobuf==5.29.3
psutil==5.9.8
pydantic==2.10.5
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.1
PyQt5==5.15.10
pyspark==3.5.4
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.2.1
requests==2.31.0
rich==13.9.4
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.4.0
shap==0.46.0
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.4
spotipy==2.25.0
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
terminado==0.18.1
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.48.1
typer==0.15.1
types-python-dateutil==2.9.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wordcloud==1.9.4
wrapt==1.17.2
xgboost==2.1.3